requirements.txt
ExifRead==2.1.2
python-dateutil==2.6.1
six==1.11.0
wheel==0.24.0
